Get more features added
Added ability to actually search for a match with
a search form. This can be of the form of a match ID
or just use a summoner name to get their most
recent ranked match.

Also added some basic map control visualization.
This is pretty rough for many reasons. See if we
can improve it given the data we have.